user-board-comment 연관관계...user엔티티는 연관관계끊어야 할듯 중간커밋/ 동작안됨

+++ src/main/java/com/sparta/lv3_board/service/BoardService.java
@@ -1,14 +1,21 @@
 package com.sparta.lv3_board.service;
 
 
-
 import com.sparta.lv3_board.dto.BoardRequestDto;
 import com.sparta.lv3_board.dto.BoardResponseDto;
 import com.sparta.lv3_board.entity.Board;
+import com.sparta.lv3_board.entity.User;
+import com.sparta.lv3_board.jwt.JwtUtil;
 import com.sparta.lv3_board.repository.BoardRepository;
+import com.sparta.lv3_board.repository.UserRepository;
+import io.jsonwebtoken.Claims;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,12 +26,33 @@ import java.util.List;
 public class BoardService {
 
     private final BoardRepository boardRepository;
+    private final UserRepository userRepository;
+    private final JwtUtil jwtUtil;
 
     //create
     public BoardResponseDto createBoard(BoardRequestDto requestDto, HttpServletRequest httpServletRequest) {
-        Board board = new Board(requestDto);
-        Board saveBoard = boardRepository.save(board);
-        return new BoardResponseDto(saveBoard);
+        // 현재 사용자 정보 가지고 오기
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+
+        String token = jwtUtil.resolveToken(httpServletRequest);
+        Claims claims;
+        // 토큰이 있는 경우에만 글 작성 가능
+        if (token != null) {
+            if (jwtUtil.validateToken(token)) {
+                claims = jwtUtil.getUserInfoFromToken(token);
+            } else {
+                throw new IllegalArgumentException("토큰 오류");
+            }
+            User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
+                    () -> new IllegalArgumentException("로그인 하세요."));
+
+            Board board = new Board(requestDto, user.getUsername());
+            Board saveBoard = boardRepository.save(board);
+
+            return new BoardResponseDto(saveBoard);
+        }
+        return null;
     }
 
 
@@ -38,6 +66,7 @@ public class BoardService {
     //read
     public Board getBoard(Long id) {
         Board board = findBoard(id);
+
         return board;
     }
 
@@ -46,15 +75,55 @@ public class BoardService {
     @Transactional
     public ResponseEntity<String> updateBoard(Long id, BoardRequestDto requestDto, HttpServletRequest httpServletRequest) {
         Board board = findBoard(id);
-        board.update(requestDto);
-        return ResponseEntity.ok("수정 성공!");
+
+        String token = jwtUtil.resolveToken(httpServletRequest);
+        Claims claims;
+        // 토큰이 있는 경우에만 글 작성 가능
+        if (token != null) {
+            if (jwtUtil.validateToken(token)) {
+                claims = jwtUtil.getUserInfoFromToken(token);
+            } else {
+                throw new IllegalArgumentException("토큰 오류");
+            }
+            User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
+                    () -> new IllegalArgumentException("로그인 하세요."));
+
+
+            if (user.getUsername().equals(board.getUsername())) {
+                board.update(requestDto);
+                return ResponseEntity.ok("수정 성공!");
+            } else {
+                throw new IllegalArgumentException("작성자가 다릅니다.");
+            }
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인이 필요합니다.");
     }
+
 
     //delete
     public ResponseEntity<String> deleteBoard(Long id, HttpServletRequest httpServletRequest) {
         Board board = findBoard(id);
-        boardRepository.delete(board);
-        return ResponseEntity.ok("삭제 성공!");
+
+        String token = jwtUtil.resolveToken(httpServletRequest);
+        Claims claims;
+
+        if (token != null) {
+            if (jwtUtil.validateToken(token)) {
+                claims = jwtUtil.getUserInfoFromToken(token);
+            } else {
+                throw new IllegalArgumentException("토큰 오류");
+            }
+            User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
+                    () -> new IllegalArgumentException("로그인 하세요."));
+
+            if (user.getUsername().equals(board.getUsername())) {
+                boardRepository.delete(board);
+                return ResponseEntity.ok("삭제 성공!");
+            } else {
+                throw new IllegalArgumentException("작성자가 다릅니다.");
+            }
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인이 필요합니다.");
     }
 
 
@@ -64,7 +133,4 @@ public class BoardService {
                 () -> new IllegalArgumentException("선택한 게시물은 존재하지 않습니다.")
         );
     }
-
-
-
 }
